refactor(notifier): remove AddReceivers function from notify.Notifier interface
AddReceivers was added to the notify.Notifier interface in ff62ba90b49ebbd91d0551080a7dea08f460e210
and the fact that it shouldn't be there was obviously overlooked during
the code review.
The AddReceivers function is present in all services so far, but it does
not define the behavior of our services. The behavior is defined only by
the Send function. The AddReceivers function is also not called once in
the code on a notify.notifier. This can also be seen by the fact that I
can remove the AddReceivers function from the notify.Notifier interface
without getting any errors in the rest of the code.