ci(docs): drop changelog generating action
I sadly don't have the time to fiddle around with this action. Even
though I'd love it to be running and auto-generating changelogs, I want
to use my spare time on notify efficiently and I don't think that this
task here has priority right now. I might be coming back to this later.

For now, changelogs must be generated manually.