feat(notify): add notifier options
Introducing Option functions. Options are meant to provide an extendable
way of instanciating a Notifier with a set of options. Currently, only
Enable()/Disable() are supported.

WithOptions() alters an already instanciated Notifier and
NewWithOptions() returns a new Notifier instance that was initialized
with the given options.

Options are passed in as a variadic list, thus, they're "optional".. no
pun intended. Calling New() results in an internal call to
NewWithOptions() without options passed, which again results in a
default Notifier being returned.