feat(notify): Add NewWithServices() constructor function
AddNewWithServices() accepts a variadic list of services and returns a
new, by New() created, Notify instance with the list of services set as
its notifiers.

If no services are given it's the functionally identical to just calling
New().

Calling NewWithServices() with a list of services is functionally equal
to calling New() and then UseServices().