refactor(sec): add ReadHeaderTimeout to verification server
WaitForOneOffVerification now uses a ReadHeaderTimeout of 20 seconds by
default. This is to avoid potential slow loris attacks which means
blocking a connection for too long. In case that the user wants to
provide a custom timeout, I added the
WaitForOneOffVerificationWithServer method. It works identical to
WaitForOneOffVerification with the subtle difference that the user can
pass in a fully customized http.Server instance.

This closes gosec G112.